managing the last item of the list
