+++ src/components/display/fewlines/LoginsOverview/LoginsOverview.tsx
@@ -2,8 +2,6 @@ import React from "react";
 import styled from "styled-components";
 
 import { IdentitySection } from "../IdentitySection/IdentitySection";
-import { NeutralLink } from "../NeutralLink";
-import type { Identity } from "@lib/@types";
 import { IdentityTypes } from "@lib/@types/Identity";
 import { SortedIdentities } from "@src/@types/SortedIdentities";
 
@@ -43,59 +41,25 @@ export const LoginsOverview: React.FC<LoginsOverviewProps> = ({
     },
   };
 
+  const identitySectionList = Object.entries(IDENTITY_SECTION_CONTENT);
+
   return (
     <Container>
-      {Object.entries(IDENTITY_SECTION_CONTENT).map(
-        ([sectionName, content]) => {
-          return <IdentitySection key={sectionName} content={content} />;
-        },
-      )}
+      {identitySectionList.map(([sectionName, content], index) => {
+        const lastOfTheList = index === identitySectionList.length - 1;
+
+        return (
+          <IdentitySection
+            key={sectionName}
+            content={content}
+            lastOfTheList={lastOfTheList}
+          />
+        );
+      })}
     </Container>
   );
 };
 
 const Container = styled.div`
   margin: 0 0 10rem 0;
-`;
-
-// const LastIdentitySections = styled(IdentitySections)`
-//   padding: 0;
-// `;
-
-export const NoIdentitiesBox = styled.p`
-  display: flex;
-  align-items: center;
-  height: 7.2rem;
-  margin-right: 0.5rem;
-  padding: 0 2rem;
-`;
-
-type BoxedLinkProps = Pick<Identity, "primary" | "status"> & {
-  social?: boolean;
-};
-
-export const BoxedLink = styled(NeutralLink)<BoxedLinkProps>`
-  height: 7.2rem;
-  padding: 0 ${({ theme }) => theme.spaces.xs};
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  ${(props) =>
-    props.social &&
-    `
-      cursor: default;
-    `}
-
-  ${(props) =>
-    props.primary &&
-    `
-        font-weight: ${props.theme.fontWeights.semibold};
-    `}
-
-  ${(props) =>
-    props.status === "unvalidated" &&
-    `
-    color: ${props.theme.colors.lightGrey};
-    `};
 `;
